Preserve any handles in syntax fallbacks

+++ tests/XFAIL/type_assertions/main.rs
@@ -136,13 +136,13 @@ impl Shape for Rectangle {
 
 impl Circle {
     pub fn area(&self) -> Rc<RefCell<Option<f64>>> {
-        return Rc::new(RefCell::new(Some(3.14159 * (*self.radius.borrow().as_ref().unwrap()) * (*self.radius.borrow().as_ref().unwrap()))));
+        return Rc::new(RefCell::new(Some(3.14159 as f64 * (*self.radius.borrow().as_ref().unwrap()) * (*self.radius.borrow().as_ref().unwrap()))));
     }
 }
 
 impl Shape for Circle {
     fn area(&self) -> Rc<RefCell<Option<f64>>> {
-        return Rc::new(RefCell::new(Some(3.14159 * (*self.radius.borrow().as_ref().unwrap()) * (*self.radius.borrow().as_ref().unwrap()))));
+        return Rc::new(RefCell::new(Some(3.14159 as f64 * (*self.radius.borrow().as_ref().unwrap()) * (*self.radius.borrow().as_ref().unwrap()))));
     }
     fn __go_clone_box(&self) -> Box<dyn Shape> {
         Box::new(self.clone()) as Box<dyn Shape>
@@ -301,7 +301,7 @@ fn main() {
     assert_without_check(Rc::new(RefCell::new(Some(Box::new(123) as Box<dyn Any>))));
 
     println!("{}", format!("{}", "\n=== Interface type assertions ===".to_string()));
-    let mut shapes = Rc::new(RefCell::new(Some(vec![Box::new(Rectangle { width: Rc::new(RefCell::new(Some(10.0))), height: Rc::new(RefCell::new(Some(5.0))), ..Default::default() }) as Box<dyn Shape>, Box::new(Circle { radius: Rc::new(RefCell::new(Some(3.0))), ..Default::default() }) as Box<dyn Shape>])));
+    let mut shapes = Rc::new(RefCell::new(Some(vec![Box::new(Rectangle { width: Rc::new(RefCell::new(Some(10.0 as f64))), height: Rc::new(RefCell::new(Some(5.0 as f64))), ..Default::default() }) as Box<dyn Shape>, Box::new(Circle { radius: Rc::new(RefCell::new(Some(3.0 as f64))), ..Default::default() }) as Box<dyn Shape>])));
 
     { let __range_holder = shapes.clone(); let __range_guard = __range_holder.borrow(); let __range_values = __range_guard.as_ref().cloned().unwrap_or_default(); drop(__range_guard); for shape in __range_values.iter() {
         describe_shape(shape.as_ref());
