+++ tests/concurrent_address_composite_method/main.rs
@@ -39,7 +39,7 @@ impl finder {
 
 pub fn call_find() -> Arc<Mutex<Option<i32>>> {
 
-    return { let __recv = (Arc::new(Mutex::new(Some(finder { base: Arc::new(Mutex::new(Some(2))), ..Default::default() })))); let __result = (*__recv.lock().unwrap().as_ref().unwrap()).find(Arc::new(Mutex::new(Some(3)))); __result };
+    return { let __recv = (Arc::new(Mutex::new(Some(finder { base: Arc::new(Mutex::new(Some(2 as i32))), ..Default::default() })))); let __result = (*__recv.lock().unwrap().as_ref().unwrap()).find(Arc::new(Mutex::new(Some(3)))); __result };
 }
 
 fn main() {
