+++ tests/concurrent_package_pointer_global/main.rs
@@ -214,7 +214,7 @@ pub(crate) static fallback: std::sync::LazyLock<std::sync::Arc<std::sync::Mutex<
 fn __go_init_globals() {
     *current.lock().unwrap() = Some(Default::default());
     *fallback.lock().unwrap() = Some(Default::default());
-    *fallback.lock().unwrap() = Some(Arc::new(Mutex::new(Some(counter { value: Arc::new(Mutex::new(Some(5))), ..Default::default() }))));
+    *fallback.lock().unwrap() = Some(Arc::new(Mutex::new(Some(counter { value: Arc::new(Mutex::new(Some(5 as i32))), ..Default::default() }))));
 }
