+++ tests/concurrent_pointer_map_keys/main.rs
@@ -59,7 +59,7 @@ fn main() {
     std::thread::spawn(move || {
         ;
     });
-    let mut n = Arc::new(Mutex::new(Some(node { value: Arc::new(Mutex::new(Some(7))), ..Default::default() })));
+    let mut n = Arc::new(Mutex::new(Some(node { value: Arc::new(Mutex::new(Some(7 as i32))), ..Default::default() })));
     let mut seen = Arc::new(Mutex::new(Some(BTreeMap::<GoLocalPtrKey<node>, Arc<Mutex<Option<i32>>>>::from([(GoLocalPtrKey::new(n.clone()), Arc::new(Mutex::new(Some(3))))]))));
     let mut current = Arc::new(Mutex::new(Some({ let __map = { let __map_holder = seen.clone(); let __map_guard = __map_holder.lock().unwrap(); let __cloned = (*__map_guard.as_ref().unwrap()).clone(); drop(__map_guard); __cloned }; __map.get(&GoLocalPtrKey::new(n.clone())).map(|__v| __v.lock().unwrap().as_ref().unwrap().clone()).unwrap_or_else(|| 0) })));
     { let __map_key = GoLocalPtrKey::new(n.clone()); let __map_value = Arc::new(Mutex::new(Some({ let __tmp_x = { let __v = (*current.lock().unwrap().as_ref().unwrap()).clone(); __v }; let __tmp_y = (*{ let __field = (*n.lock().unwrap().as_ref().unwrap()).value.clone(); __field }.lock().unwrap().as_ref().unwrap()); __tmp_x + __tmp_y }))); (*seen.lock().unwrap().as_mut().unwrap()).insert(__map_key, __map_value); };
