+++ tests/error_simple/main.rs
@@ -5,7 +5,7 @@ use std::rc::{Rc};
 pub fn divide(a: Rc<RefCell<Option<i32>>>, b: Rc<RefCell<Option<i32>>>) -> (Rc<RefCell<Option<i32>>>, Rc<RefCell<Option<Box<dyn StdError>>>>) {
 
     if (*b.borrow().as_ref().unwrap()) == 0 {
-        return (Rc::new(RefCell::new(Some(0))), Rc::new(RefCell::new(Some(Box::<dyn std::error::Error>::from("division by zero".to_string())))));
+        return (Rc::new(RefCell::new(Some(0 as i32))), Rc::new(RefCell::new(Some(Box::<dyn std::error::Error>::from("division by zero".to_string())))));
     }
     return ({
             let __tmp_x = (*a.borrow().as_ref().unwrap());
